+++ src/main/java/me/flotsam/frettler/Main.java
@@ -10,29 +10,23 @@ import me.flotsam.frettler.command.MandolinCommand;
 import me.flotsam.frettler.command.UkeleleCommand;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
+import picocli.CommandLine.HelpCommand;
 
-@Command(
-    name = "frettler", 
-    description = "Generates guitar scales and chords",
-    subcommands = {
-        CompletionsCommand.class,
-        GuitarCommand.class,
-        BassGuitarCommand.class,
-        BanjoCommand.class,
-        MandolinCommand.class,
-        UkeleleCommand.class
-    }
-)
+@Command(name = "frettler",
+    description = "Generates scalei/nodes, chords and arpeggios for fretted instruments",
+    subcommands = {HelpCommand.class, CompletionsCommand.class, GuitarCommand.class,
+        BassGuitarCommand.class, BanjoCommand.class, MandolinCommand.class, UkeleleCommand.class})
 public class Main implements Callable<Integer> {
 
   public static void main(String... args) throws Exception {
-    new CommandLine(new Main()).setCaseInsensitiveEnumValuesAllowed(true).execute(args);
+    new CommandLine(new Main()).setCaseInsensitiveEnumValuesAllowed(true)
+        .setAbbreviatedOptionsAllowed(true).setAbbreviatedSubcommandsAllowed(true).execute(args);
   }
 
   @Override
   public Integer call() throws Exception {
-      CommandLine.usage(new FrettlerCommand(), System.out);
-      return 0;
+    CommandLine.usage(new FrettlerCommand(), System.out);
+    return 0;
   }
 }