+++ src/main/java/me/flotsam/frettler/command/GuitarCommand.java
@@ -14,6 +14,6 @@ public class GuitarCommand extends FrettedInstrumentCommand implements Runnable 
   
   @Override
   public void run() {
-   exec(new Guitar(strings));
+   exec(new Guitar(strings, frets));
   }
 }