+++ src/main/java/me/flotsam/frettler/instrument/Banjo.java
@@ -9,11 +9,11 @@ public class Banjo extends FrettedInstrument {
   private static final Note[] DEFAULT_STRINGS = new Note[] {Note.G, Note.D, Note.G, Note.B, Note.D};
 
   public Banjo() {
-    this(DEFAULT_STRINGS);
+    this(DEFAULT_STRINGS, FRETS);
   }
 
-  public Banjo(Note[] strings) {
-    super("Banjo", FRETS, strings.length > 0 ? strings : DEFAULT_STRINGS);
+  public Banjo(Note[] strings, Integer frets) {
+    super("Banjo", frets, strings.length > 0 ? strings : DEFAULT_STRINGS);
   }
   
   public boolean isBanjo() {