+++ src/main/java/me/flotsam/frettler/view/HorizontalView.java
@@ -1,6 +1,7 @@
 package me.flotsam.frettler.view;
 
 import static java.lang.System.out;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -14,57 +15,51 @@ import me.flotsam.frettler.engine.ScaleNote;
 import me.flotsam.frettler.instrument.Fret;
 import me.flotsam.frettler.instrument.FrettedInstrument;
 
-public class FretboardView {
+public class HorizontalView {
 
+  private static final List<Integer> inlays = Arrays.asList(1, 3, 5, 7, 9, 12, 15, 17, 19, 21, 23);
   private FrettedInstrument instrument;
   private Options defaultOptions = new Options(false, true, true);
 
-  public FretboardView(FrettedInstrument instrument) {
+  public HorizontalView(FrettedInstrument instrument) {
     this.instrument = instrument;
   }
 
   public void showFretboard() {
-    showFretboard(Scale.CHROMATIC_SCALE, defaultOptions);
+    showScale(Scale.CHROMATIC_SCALE, defaultOptions);
   }
 
   public void showFretboard(Options options) {
-    showFretboard(Scale.CHROMATIC_SCALE, options);
+    showScale(Scale.CHROMATIC_SCALE, options);
   }
 
-  public void showFretboard(Chord chord) {
-    showFretboard(chord, defaultOptions);
+  public void showChord(Chord chord) {
+    showChord(chord, defaultOptions);
   }
 
-  public void showFretboard(Chord chord, Options options) {
+  public void showChord(Chord chord, Options options) {
     out.println();
     out.print("    ");
     out.println(StringUtils.center(chord.getTitle() + " ~ (" + instrument.getLabel() + " [" + instrument.getStringNotes().stream().map(Note::name).collect(Collectors.joining(",")) + "]" , 84));
     out.println();
-    showFretboard(chord.getChordNotes(), options);
+    display(chord.getChordNotes(), options);
   }
 
-  public void showFretboard(Scale scale) {
-    showFretboard(scale, defaultOptions);
+  public void showScale(Scale scale) {
+    showScale(scale, defaultOptions);
   }
 
-  public void showFretboard(Scale scale, Options options) {
+  public void showScale(Scale scale, Options options) {
     out.println();
     out.print("    ");
     out.println(StringUtils.center(scale.getTitle() + " ~ (" + instrument.getLabel() + " [" + instrument.getStringNotes().stream().map(Note::name).collect(Collectors.joining(",")) + "]" , 84));
     out.println();
-    showFretboard(scale.getScaleNotes(), options);
+    display(scale.getScaleNotes(), options);
   }
 
-  public void showFretboard(List<ScaleNote> scaleNotes, Options options) {
+  public void display(List<ScaleNote> scaleNotes, Options options) {
     ColourMap cm = new ColourMap();
-    if (options.isInlays()) {
-      out.println(
-          "   ┏━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━━━━━━━━.━━━┓");
-    } else {
-      out.println(
-          "   ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓");
-    }
-
+    out.println(createFretboardSide(true, options));
 
     for (int i = instrument.getFretsByString().size() - 1; i >= 0; i--) {
       List<Fret> tonesInString = instrument.getFretsByString().get(i);
@@ -78,7 +73,7 @@ public class FretboardView {
           if (options.isIntervals()) {
             fretStr = note.get().getInterval().isPresent() ? note.get().getInterval().get().getLabel() : note.get().getNote().getLabel();
           } else {
-              fretStr = note.get().getNote().getLabel();
+            fretStr = note.get().getNote().getLabel();
           }
           if (options.isColour()) {
             Colour col = cm.get(fretStr);
@@ -117,18 +112,30 @@ public class FretboardView {
       }
       out.println(stringBuilder.toString());
     }
-
-    if (options.isInlays()) {
-      out.println(
-          "   └━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━.━━━━━━━━━━━━━━━━━━━━.━━━┛");
-    } else {
-      out.println(
-          "   └━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛");
-    }
+    
+    out.println(createFretboardSide(false, options));
     out.println();
     out.println();
   }
   
+  private String createFretboardSide(boolean upper, Options options) {
+   String lowerTop = "┗";
+   String lowerBottom = "┛";
+   String upperTop = "┏";
+   String upperBottom = "┓";
+   String side = "━";
+   String inlay = ".";
+    
+   StringBuilder sb = new StringBuilder("   ").append(upper ? upperTop : lowerTop);
+   for (int fret = 1; fret <= instrument.getFrets(); fret++) {
+    sb.append(side.repeat(2));
+    sb.append(options.isInlays() && inlays.contains(fret) ? inlay : side);
+    sb.append(side.repeat(fret == instrument.getFrets() ? 3 : 4));
+   }
+   sb.append(upper ? upperBottom : lowerBottom);
+   
+   return sb.toString();
+  }
 
   @Data
   @AllArgsConstructor